REF: Rename exchange -> interchange
