+++ pandas/core/frame.py
@@ -220,8 +220,8 @@ import pandas.plotting
 
 if TYPE_CHECKING:
 
-    from pandas.core.exchange.dataframe_protocol import DataFrame as DataFrameXchg
     from pandas.core.groupby.generic import DataFrameGroupBy
+    from pandas.core.interchange.dataframe_protocol import DataFrame as DataFrameXchg
     from pandas.core.internals import SingleDataManager
     from pandas.core.resample import Resampler
 
@@ -819,7 +819,7 @@ class DataFrame(NDFrame, OpsMixin):
         self, nan_as_null: bool = False, allow_copy: bool = True
     ) -> DataFrameXchg:
         """
-        Return the dataframe exchange object implementing the exchange protocol.
+        Return the dataframe interchange object implementing the interchange protocol.
 
         Parameters
         ----------
@@ -832,19 +832,19 @@ class DataFrame(NDFrame, OpsMixin):
 
         Returns
         -------
-        DataFrame exchange object
+        DataFrame interchange object
             The object which consuming library can use to ingress the dataframe.
 
         Notes
         -----
-        Details on the exchange protocol:
+        Details on the interchange protocol:
         https://data-apis.org/dataframe-protocol/latest/index.html
 
         `nan_as_null` currently has no effect; once support for nullable extension
         dtypes is added, this value should be propagated to columns.
         """
 
-        from pandas.core.exchange.dataframe import PandasDataFrameXchg
+        from pandas.core.interchange.dataframe import PandasDataFrameXchg
 
         return PandasDataFrameXchg(self, nan_as_null, allow_copy)
