+++ pandas/core/interchange/column.py
@@ -12,14 +12,14 @@ from pandas.api.types import (
     is_categorical_dtype,
     is_string_dtype,
 )
-from pandas.core.exchange.buffer import PandasBuffer
-from pandas.core.exchange.dataframe_protocol import (
+from pandas.core.interchange.buffer import PandasBuffer
+from pandas.core.interchange.dataframe_protocol import (
     Column,
     ColumnBuffers,
     ColumnNullType,
     DtypeKind,
 )
-from pandas.core.exchange.utils import (
+from pandas.core.interchange.utils import (
     ArrowCTypes,
     Endianness,
     NoBufferPresent,
@@ -136,7 +136,7 @@ class PandasColumn(Column):
         kind = _NP_KINDS.get(dtype.kind, None)
         if kind is None:
             # Not a NumPy dtype. Check if it's a categorical maybe
-            raise ValueError(f"Data type {dtype} not supported by exchange protocol")
+            raise ValueError(f"Data type {dtype} not supported by interchange protocol")
 
         return kind, dtype.itemsize * 8, dtype_to_arrow_c_fmt(dtype), dtype.byteorder
