+++ pandas/core/interchange/dataframe.py
@@ -4,8 +4,8 @@ from collections import abc
 from typing import TYPE_CHECKING
 
 import pandas as pd
-from pandas.core.exchange.column import PandasColumn
-from pandas.core.exchange.dataframe_protocol import DataFrame as DataFrameXchg
+from pandas.core.interchange.column import PandasColumn
+from pandas.core.interchange.dataframe_protocol import DataFrame as DataFrameXchg
 
 if TYPE_CHECKING:
     from pandas import Index
