+++ pandas/core/interchange/dataframe_protocol.py
@@ -389,7 +389,7 @@ class DataFrame(ABC):
 
     @abstractmethod
     def __dataframe__(self, nan_as_null: bool = False, allow_copy: bool = True):
-        """Construct a new exchange object, potentially changing the parameters."""
+        """Construct a new interchange object, potentially changing the parameters."""
         pass
 
     @property
