+++ pandas/core/interchange/utils.py
@@ -1,5 +1,5 @@
 """
-Utility functions and objects for implementing the exchange API.
+Utility functions and objects for implementing the interchange API.
 """
 
 from __future__ import annotations
